Move orm, reducers, store and events to core folder
