+++ app/components/Login.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { forEach } from 'lodash';
 import './Login.global.css';
-import Events from '../events';
+import Events from '../core/events';
 import NotificationsQueue from './NotificationsQueue';
 
 class Login extends Component {
