+++ app/components/dashboard/TabPanel.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux';
 import pretty from 'prettysize';
 import PropTypes from 'prop-types';
 import './Dashboard.global.css';
-import Events from '../../events';
-import selectors from '../../orm/selectors';
+import Events from '../../core/events';
+import selectors from '../../core/orm/selectors';
 import { ReactComponent as SuccessIcon } from '../../icons/svg/successFilled.svg';
 import { ReactComponent as WarningIcon } from '../../icons/svg/warningFilled.svg';
 import { ReactComponent as PlayIcon } from '../../icons/svg/playFilled.svg';
