+++ app/core/reducers/dashboard.js
@@ -1,5 +1,5 @@
 import { handleAction } from 'redux-actions';
-import { GET_DASHBOARD } from '../core/actions/dashboard';
+import { GET_DASHBOARD } from '../actions/dashboard';
 
 const dashboard = handleAction(
   GET_DASHBOARD,
