+++ app/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
 import Root from './pages/root';
-import { store, history } from './store/configureStore';
+import { store, history } from './core/store/configureStore';
 import './app.global.css';
 
 render(
